Add validation to prevent division archival while in use
This commit addresses the scenario when a "division" is in use, through 'department' checking before attempting to archive it. In addition, the routes/api.php file has been updated to include an import route for CapEx, along with the corresponding controller method (capexImport). Moreover, changes have been made into 'CapexController' for code optimisation, sorting by 'created_at' in descending order, and 'import capex' method implementation for importing CapEx's data. The commit also includes the removal of unused code throughout multiple files and an efficient approach for activating printer ip in 'PrinterIPController,' minimizing database operations (reduced code complexity and improved performance).

+++ app/Http/Controllers/Masterlist/DivisionController.php
@@ -182,6 +182,10 @@ class DivisionController extends Controller
                 if ($checkFixedAsset) {
                     return response()->json(['error' => 'Unable to archived , Division is still in use!'], 422);
                 }
+                $checkDepartment = Department::where('division_id', $id)->exists();
+                if ($checkDepartment) {
+                    return response()->json(['error' => 'Unable to archived , Division is still in use!'], 422);
+                }
 
                 $updateStatus = $Division->where('id', $id)->update(['is_active' => false]);
                 $Division->where('id', $id)->delete();
